+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+**2025-03-01**
+
+- Add `goimports`
+- Add `gofumpt`
+- Add go to error line with: option + G
+
 **2025-02-28**
 
 - Initial development started